Pieces being placed works, imgui looks better, animations need work
The pieces get placed in the correct locations when the board is created but they are of type 'None' so they don't get rendered.

Imgui has menus now:
1. lets you control where you will place the next piece
2. lets you quit out of the game using a button

The animation works fine now, but when I play the animation then all the checkers disappear and I have no idea why.